Renamed Apps
$admins -> administrator\ndevelopers -> developer/ncustomers -> customer
