+++ backend/users/backend/administrator/apps.py
@@ -1,0 +1,6 @@
+from common.apps import BaseAppConfig
+
+
+class AdministratorAppConfig(BaseAppConfig):
+    name = "administrator"
+    verbose_name = "Administrator"
