--- backend/users/backend/admins/apps.py
@@ -1,6 +1,0 @@
-from common.apps import BaseAppConfig
-
-
-class AdminAppConfig(BaseAppConfig):
-    name = "admins"
-    verbose_name = "Admin"
