+++ backend/users/backend/developer/apps.py
@@ -2,5 +2,5 @@ from common.apps import BaseAppConfig
 
 
 class DeveloperAppConfig(BaseAppConfig):
-    name = "developers"
+    name = "developer"
     verbose_name = "Developer"
